Resolved - Qa Sheet Bug
If we delete any uploaded document and click on reset option after this when we click on update button we get ""ctrl+f5"" error and page start to buffer.
Now, if we search for that student and check for that document paramter then document does not gets changed which is expected result.